Version 2.3
Se cambio el avion que disparaba por un helicoptero(le falta la rotacion de las helices)
Se agregaron botones de pausa y para volver al main menu o a los mapas para escoger entre otras cosas

+++ core/src/tools/CollisionRect.java
@@ -2,17 +2,17 @@ package tools;
 
 public class CollisionRect {
 
-    int x,y;
+    float x,y;
     int width, height;
 
-     public CollisionRect(int x, int y, int width, int height){
+     public CollisionRect(float x, float y, int width, int height){
          this.x = x;
          this.y = y;
          this.width = width;
          this.height = height;
      }
 
-     public void move (int x, int y){
+     public void move (float x, float y){
          this.x = x;
          this.y = y;
      }
